Refactored oxen df --sql to reasonable format
